requirements.txt
asgiref==3.8.1
decouple==0.0.7
Django==5.1.4
flake8==7.1.1
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.5.3
